break up the monolith
